refactor for class factory for events
Revamped the way the different events are treated where they now each
implement the function Execute() and can be abstracted to an abstract
type Event.
The events are also now responsible for their own parsing of the json
event by each exposing a New() method.

Visit is still responsible for sending the actual UA events because it's
the container for all the information on LastQuery, LastResponse etc.

Deleted and re-added visit.go to the package so unfortunately lost track
of the changes line by line.